+++ linearmodels/linear.go
@@ -16,7 +16,7 @@ type LinearRegressionModel struct {
     inputDim   int
 }
 
-func (m *LinearRegressionModel) Fit(input [][]float64, target []float64, maxIterations int) disgo.Parameters {
+func (m *LinearRegressionModel) Fit(input [][]float64, target []float64, maxIterations uint64) disgo.Parameters {
 
     if m.Bias { input = AddBiasTermToInputTable(input) }
